Changed how dependencies are versioned.
First change: `package.json` now only tracks the major (if `^`) or minor (if `~`) versions, and serves as a human-manageable declaration of intent.

Second change: `package-lock.json`, which is generated automatically (sometimes inconveniently -- it's not uncommon for people to accidentally commit it in commits unrelated to dependency updates), is changed to `npm-shrinkwrap.json`, which is manually generated by the `npm shrinkwrap` command.

Third change: `package-lock.json`/`npm-shrinkwrap.json` are now generated in a minified format, and `git` is told to treat them as binaries instead of diff'able text.

Fourth change: I ran `npm update` and `npm audit fix`.