Added additional settings to `.npmrc`
`preid` is empty by default, but the `npm` documentation suggests eg "rc".

`save-prefix` here is the same as the default;  but it's handy to have it explicitly set in your `.npmrc` -- partially so that it's clear, and partially so that you can easily change to to, eg, `'~'` if you want.

`loglevel` is in a similar situation:  `'notice'` is the default;  but I've enumerated it here so that it is clear, and so that it can be easily changed later -- eg, to `'warn'`.

`logs-max` defaults to `10`, which imho is much too high.  I've here limited it to the human subitization range:  `4`.

`access` is set to `restricted` here so that even unscoped packages cannot be accidentally published to the npmjs registry.  Shouldn't be a concern anyway, since `"private"` is set to `true` in `package.json`.